Move free variables code to a separate file
